php-migration-tool/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pyyaml==6.0.1
jinja2==3.1.2
openai==1.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
gitpython==3.1.40
requests==2.31.0
httpx==0.25.2
